+++ backend/src/shared/constants.ts
@@ -1,6 +1,6 @@
 import { CookieOptions } from 'express';
 
-import { mode, Mode } from '../../environment';
+import { mode, Mode } from '../environment';
 
 const isInProd = mode === Mode.Production;
 
@@ -13,3 +13,5 @@ export const userIdCookieOptions: CookieOptions = {
 };
 
 export const userIdCookieKey = 'USER_ID';
+
+export const unauthenticatedStatus = 'client_is_not_authorized';
